Added CurveBall component
CurveBall slightly increments the angle of the ball while it's
overlapping.

collision.ball_circle now only modifies state is update=True
The electric sweep is a cool noise. Consider putting it in.
lose1.mp3 is now louder.

The two big things to do now are add the coin class and fill in gaps.